feat(startScreen): Refactored all of the start screen to keep everything detached to get ready for gui changes for the room screen and lobby into the game.
